requirements.txt
discord.py
openai
flask
requests
httpx
